1. Project Refactor. 2. Google Sign In Project Setup in API Console 3. Google Sign In Pod Installed
