--- Calendar/AppDelegate.swift
@@ -1,22 +1,0 @@
-//
-//  AppDelegate.swift
-//  Calendar
-//
-//  Created by Abhijeet Choudhary on 06/08/22.
-//
-
-import UIKit
-
-@main
-class AppDelegate: UIResponder, UIApplicationDelegate {
-    
-    var window: UIWindow?
-    
-    func application(_ application: UIApplication,
-                     didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        // Override point for customization after application launch.
-        return true
-    }
-    
-}
-
