+++ Calendar/Resources/Assets.xcassets/Contents.json
