--- Calendar/ViewController.swift
@@ -1,25 +1,0 @@
-//
-//  ViewController.swift
-//  Calendar
-//
-//  Created by Abhijeet Choudhary on 06/08/22.
-//
-
-import UIKit
-
-final class ViewController: UIViewController {
-    
-    @IBOutlet private weak var loginButton: UIButton!
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-    
-    @IBAction private func loginButtonTapped(_ sender: UIButton) {
-        
-    }
-    
-    
-}
-
